xsps_main.c: use size_t to make it build everywhere.

+++ src/xsps_main.c
@@ -12,7 +12,7 @@ int main(int argc, char** argv) {
 	xsps_arg_parse(xhp, argc, argv);
 
 	xsps_log_info(xhp, "%s -- %s!!!", "test", "hello");
-	xsps_log_warn(xhp, "%c %s %x %d", '!', "hi", (unsigned int)&main, 50);
+	xsps_log_warn(xhp, "%c %s %x %d", '!', "hi", (size_t)&main, 50);
 
 	xsps_handle_free(xhp);
 	return 0;
